// src/models/ECountries.ts
export enum ECountries {
    // Европа
    ARMENIA = 'Armenia',
    RUSSIA = 'Russia',
    EUROPA = 'EUROPA',
    GERMANY = 'Germany',
    FRANCE = 'France',
    ITALY = 'Italy',
    SPAIN = 'Spain',
    UKRAINE = 'Ukraine',
    POLAND = 'Poland',
    NETHERLANDS = 'Netherlands',
    BELGIUM = 'Belgium',
    SWITZERLAND = 'Switzerland',
    SWEDEN = 'Sweden',
    NORWAY = 'Norway',
    FINLAND = 'Finland',
    DENMARK = 'Denmark',
    CZECH_REPUBLIC = 'Czech Republic',
    AUSTRIA = 'Austria',
    GREECE = 'Greece',
    PORTUGAL = 'Portugal',
    HUNGARY = 'Hungary',
    ROMANIA = 'Romania',
    BULGARIA = 'Bulgaria',
    SLOVAKIA = 'Slovakia',
    CROATIA = 'Croatia',
    SERBIA = 'Serbia',
    LITHUANIA = 'Lithuania',
    LATVIA = 'Latvia',
    ESTONIA = 'Estonia',
    ICELAND = 'Iceland',
    LUXEMBOURG = 'Luxembourg',
    IRELAND = 'Ireland',
    UNITED_KINGDOM = 'United Kingdom',
    BOSNIA_AND_HERZEGOVINA = 'Bosnia and Herzegovina',
    ALBANIA = 'Albania',
    MOLDOVA = 'Moldova',
    NORTH_MACEDONIA = 'North Macedonia',
    MONTENEGRO = 'Montenegro',
    BELARUS = 'Belarus',
    MALTA = 'Malta',
  
    // Америка
    AMERICA = 'America',
    UNITED_STATES = 'United States',
    CANADA = 'Canada',
    MEXICO = 'Mexico',
    BRAZIL = 'Brazil',
    ARGENTINA = 'Argentina',
    CHILE = 'Chile',
    COLOMBIA = 'Colombia',
    PERU = 'Peru',
    VENEZUELA = 'Venezuela',
    CUBA = 'Cuba',
    ECUADOR = 'Ecuador',
    PARAGUAY = 'Paraguay',
    URUGUAY = 'Uruguay',
    BOLIVIA = 'Bolivia',
    PANAMA = 'Panama',
    JAMAICA = 'Jamaica',
    GUATEMALA = 'Guatemala',
  
    // Азия
    CHINA = 'China',
    JAPAN = 'Japan',
    SOUTH_KOREA = 'South Korea',
    INDIA = 'India',
    PAKISTAN = 'Pakistan',
    BANGLADESH = 'Bangladesh',
    VIETNAM = 'Vietnam',
    INDONESIA = 'Indonesia',
    MALAYSIA = 'Malaysia',
    PHILIPPINES = 'Philippines',
    THAILAND = 'Thailand',
    SINGAPORE = 'Singapore',
    SAUDI_ARABIA = 'Saudi Arabia',
    UNITED_ARAB_EMIRATES = 'United Arab Emirates',
    IRAN = 'Iran',
    IRAQ = 'Iraq',
    TURKEY = 'Turkey',
    ISRAEL = 'Israel',
    JORDAN = 'Jordan',
    LEBANON = 'Lebanon',
    SYRIA = 'Syria',
    SRI_LANKA = 'Sri Lanka',
    KAZAKHSTAN = 'Kazakhstan',
    UZBEKISTAN = 'Uzbekistan',
    KYRGYZSTAN = 'Kyrgyzstan',
    TAJIKISTAN = 'Tajikistan',
    TURKMENISTAN = 'Turkmenistan',
    GEORGIA = 'Georgia',
    AZERBAIJAN = 'Azerbaijan',
  
    // Африка
    SOUTH_AFRICA = 'South Africa',
    EGYPT = 'Egypt',
    NIGERIA = 'Nigeria',
    KENYA = 'Kenya',
    ETHIOPIA = 'Ethiopia',
    MOROCCO = 'Morocco',
    ALGERIA = 'Algeria',
    TUNISIA = 'Tunisia',
    GHANA = 'Ghana',
    TANZANIA = 'Tanzania',
    UGANDA = 'Uganda',
    ZAMBIA = 'Zambia',
    ZIMBABWE = 'Zimbabwe',
  
    // Океания
    AUSTRALIA = 'Australia',
    NEW_ZEALAND = 'New Zealand',
    FIJI = 'Fiji',
    PAPUA_NEW_GUINEA = 'Papua New Guinea',
    SAMOA = 'Samoa',
  }